fix(Peek): prevent OverflowException in NumberOfDigits for int.MinValue
Cast to long before Math.Abs to avoid OverflowException when num == int.MinValue,
since |int.MinValue| > int.MaxValue.

Closes #46960

+++ src/modules/peek/Peek.Common/Helpers/MathHelper.cs
@@ -16,7 +16,7 @@ namespace Peek.Common.Helpers
 
         public static int NumberOfDigits(int num)
         {
-            return Math.Abs(num).ToString(CultureInfo.InvariantCulture).Length;
+            return Math.Abs((long)num).ToString(CultureInfo.InvariantCulture).Length;
         }
     }
 }
